+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## [0.1.2] - 2024-11-19
 ### Added
 - Support for wasm32-unknown-unknown target.
+- Implement `TrafficShaperPayload` for `Result` and `Option`.